core/requirements.txt
imageio==2.26.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
PyYAML==6.0